ex03 correction

+++ cpp02/ex03/srcs/bsp.cpp
@@ -6,16 +6,17 @@
 /*   By: brunodeoliveira <brunodeoliveira@studen    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/05 01:54:45 by brunodeoliv       #+#    #+#             */
-/*   Updated: 2022/05/05 03:07:17 by brunodeoliv      ###   ########.fr       */
+/*   Updated: 2022/05/05 23:58:09 by brunodeoliv      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Point.hpp"
+#include "Fixed.hpp"
 
 static float	sign(Point const a, Point const b, Point const c)
 {
-	return ((a.getX() - c.getX()) * (b.getY() - c.getY())
-			- (b.getX() - c.getX()) * (a.getY() - c.getY())).toFloat();
+	return (a.getX().toFloat() - c.getX().toFloat()) * (b.getY().toFloat() - c.getY().toFloat())
+			- (b.getX().toFloat() - c.getX().toFloat()) * (a.getY().toFloat() - c.getY().toFloat());
 }
 
 bool bsp( Point const a, Point const b, Point const c, Point const point)
@@ -27,7 +28,6 @@ bool bsp( Point const a, Point const b, Point const c, Point const point)
 	s1 = sign(point, a, b);
 	s2 = sign(point, b, c);
 	s3 = sign(point, c, a);
-
 	hasNeg = (s1 < 0) || (s2 < 0) || (s3 < 0);
 	hasPos = (s1 > 0) || (s2 > 0) || (s3 > 0);
